+++ src/providers/colors/githubTheme.js
@@ -1,0 +1,39 @@
+const aadTheme_GitHub = {
+
+  globalLinkColor: '#478be6',
+  feed: {
+    bg: 'var(--bgColor-default, var(--color-canvas-default))',
+  },
+  container: {
+    bg: 'var(--bgColor-default, var(--color-canvas-default))',
+    border: 'var(--borderColor-default)',
+  },
+  widget: {
+    bg: '#282f38',
+    secondary: '#222a31',
+  },
+  modal: {
+    bg: 'var(--bgColor-default, var(--color-canvas-default))',
+    'inset-bg': 'var(--bgColor-muted, var(--color-canvas-inset))'
+  },
+  
+  drawer: {
+    bg: 'var(--bgColor-default, var(--color-canvas-default))',
+    'inset-bg': 'var(--bgColor-muted, var(--color-canvas-inset))'
+  },
+  settings: {
+    icon: {
+      bg: 'transparent',
+      border: 'var(--button-default-borderColor-rest)',
+      iconFill: 'var(--fgColor-muted, var(--color-fg-muted))',
+    }
+  },
+
+
+  addNewWidget: {
+    bg: 'var(--bgColor-muted, var(--color-canvas-subtle))'
+  },
+  recentActivity: {
+    'item-border': 'var(--borderColor-default)',
+  }
+}